requirements.txt
matplotlib==3.8.2
numpy==1.26.3
Pillow==10.2.0
rasterio==1.3.9
scikit_learn==1.3.2
torch==2.1.2+cu118
torchvision==0.16.2+cu118
segmentation_models_pytorch==0.3.3
tqdm==4.65.0
